backend/payouts/exceptions.py
import logging
from rest_framework.views import exception_handler
from rest_framework.response import Response

logger = logging.getLogger(__name__)


def custom_exception_handler(exc, context):
    response = exception_handler(exc, context)

    if response is None:
        logger.exception("Unhandled exception in view: %s", exc)
        return Response({"error": "Internal server error"}, status=500)

    return response
