backend/requirements.txt
django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.3
whitenoise==6.6.0
gunicorn==21.2.0
python-dotenv==1.0.1
